+++ fastmri_recon/helpers/torch_utils.py
@@ -20,10 +20,3 @@ class ConvBlock(torch.nn.Module):
     def forward(self, x):
         y = self.overall_convs(x)
         return y
-
-def replace_values_on_mask_torch(cnn_fft, kspace, mask):
-    mask = mask[..., None]
-    mask = mask.expand_as(kspace).float()
-    anti_mask = 1.0 - mask
-    replace_cnn_fft = anti_mask * cnn_fft + kspace
-    return replace_cnn_fft
