+++ fastmri_recon/models/cascading.py
@@ -1,40 +1,13 @@
-from keras.layers import Input, Lambda, Conv2D, Add, Layer
+from keras.layers import Input, Lambda
 from keras.models import Model
-from keras.optimizers import Adam
-import tensorflow as tf
 import torch
 from torch import nn
 
-from .pdnet_crop import tf_adj_op, tf_op, concatenate_real_imag, complex_from_half, tf_crop, tf_unmasked_adj_op, tf_unmasked_op
-from ..helpers.utils import keras_psnr, keras_ssim
-from ..helpers.torch_utils import ConvBlock, replace_values_on_mask_torch
+from ..helpers.keras_utils import default_model_compile
+from ..helpers.nn_mri_helpers import tf_fastmri_format, tf_unmasked_adj_op, tf_unmasked_op, replace_values_on_mask_torch, MultiplyScalar, conv2d_complex, enforce_kspace_data_consistency
+from ..helpers.torch_utils import ConvBlock
 from ..helpers.transforms import ifft2, fft2, center_crop, complex_abs
 
-class MultiplyScalar(Layer):
-    def __init__(self, **kwargs):
-        super(MultiplyScalar, self).__init__(**kwargs)
-
-    def build(self, input_shape):
-        # Create a trainable weight variable for this layer.
-        self.sample_weight = self.add_weight(
-            name='sample_weight',
-            shape=(1,),
-            initializer='ones',
-            trainable=True,
-        )
-        super(MultiplyScalar, self).build(input_shape)  # Be sure to call this at the end
-
-    def call(self, x):
-        return tf.cast(self.sample_weight, tf.complex64) * x
-
-    def compute_output_shape(self, input_shape):
-        return input_shape
-
-def replace_values_on_mask(x):
-    cnn_fft, kspace_input, mask = x
-    anti_mask = tf.expand_dims(tf.dtypes.cast(1.0 - mask, cnn_fft.dtype), axis=-1)
-    replace_cnn_fft = tf.math.multiply(anti_mask, cnn_fft) + kspace_input
-    return replace_cnn_fft
 
 def cascade_net(input_size=(640, None, 1), n_cascade=5, n_convs=5, n_filters=16, noiseless=True, lr=1e-3):
     mask_shape = input_size[:-1]
@@ -43,52 +16,20 @@ def cascade_net(input_size=(640, None, 1), n_cascade=5, n_convs=5, n_filters=16,
 
     zero_filled = Lambda(tf_unmasked_adj_op, output_shape=input_size, name='ifft_simple')(kspace_input)
 
-
     image = zero_filled
     multiply_scalar = MultiplyScalar()
     for i in range(n_cascade):
         # residual convolution
-        res_image = concatenate_real_imag(image)
-        for j in range(n_convs):
-            res_image = Conv2D(
-                n_filters,
-                3,
-                activation='relu',
-                padding='same',
-                kernel_initializer='glorot_uniform',
-            )(res_image)
-        res_image = Conv2D(
-            2,
-            3,
-            activation='linear',
-            padding='same',
-            kernel_initializer='glorot_uniform',
-        )(res_image)
-        res_image = complex_from_half(res_image, 1, input_size)
-        image = Add(name='res_connex_{i}'.format(i=i+1))([image, res_image])
+        image = conv2d_complex(image, n_filters, n_convs, output_shape=input_size, res=True)
         # data consistency layer
-        if noiseless:
-            cnn_fft = Lambda(tf_unmasked_op, output_shape=input_size, name='fft_simple_{i}'.format(i=i+1))(image)
-            data_consistency_fourier = Lambda(replace_values_on_mask, output_shape=input_size, name='fft_repl_{i}'.format(i=i+1))([cnn_fft, kspace_input, mask])
-            image = Lambda(tf_unmasked_adj_op, output_shape=input_size, name='ifft_simple_{i}'.format(i=i+1))(data_consistency_fourier)
-        else:
-            cnn_fft = Lambda(tf_op, output_shape=input_size, name='fft_masked_{i}'.format(i=i+1))([image, mask])
-            cnn_fft = Lambda(lambda x: -x, output_shape=input_size)(cnn_fft)
-            data_consistency_fourier = Add(name='data_consist_fft_{i}'.format(i=i+1))([kspace_input, cnn_fft])
-            data_consistency_image = Lambda(tf_adj_op, output_shape=input_size, name='ifft_masked_{i}'.format(i=i+1))([data_consistency_fourier, mask])
-            data_consistency_image = multiply_scalar(data_consistency_image)
-            image = Add(name='data_consist_enf_{i}'.format(i=i+1))([image, data_consistency_image])
-
+        kspace = Lambda(tf_unmasked_op, output_shape=input_size, name='fft_simple_{i}'.format(i=i+1))(image)
+        kspace = enforce_kspace_data_consistency(kspace, kspace_input, mask, input_size, multiply_scalar, noiseless)
+        image = Lambda(tf_unmasked_adj_op, output_shape=input_size, name='ifft_simple_{i}'.format(i=i+1))(kspace)
     # module and crop of image
-    image = Lambda(tf.math.abs, name='image_module', output_shape=input_size)(image)
-    image = Lambda(tf_crop, name='cropping', output_shape=(320, 320, 1))(image)
+    image = tf_fastmri_format(image)
     model = Model(inputs=[kspace_input, mask], outputs=image)
 
-    model.compile(
-        optimizer=Adam(lr=lr, clipnorm=1.),
-        loss='mean_absolute_error',
-        metrics=['mean_squared_error', keras_psnr, keras_ssim],
-    )
+    default_model_compile(model, lr)
 
     return model
